Big filtering refactor + implementation
