+++ apps/web/ui/analytics/analytics-provider.tsx
@@ -5,8 +5,6 @@ import {
   ANALYTICS_VIEWS,
   DUB_LINKS_ANALYTICS_INTERVAL,
   DUB_PARTNERS_ANALYTICS_INTERVAL,
-  EVENT_TYPES,
-  VALID_ANALYTICS_FILTERS,
 } from "@/lib/analytics/constants";
 import {
   AnalyticsResponseOptions,
@@ -15,13 +13,11 @@ import {
   EventType,
 } from "@/lib/analytics/types";
 import { editQueryString } from "@/lib/analytics/utils";
-import { combineTagIds } from "@/lib/api/tags/combine-tag-ids";
 import usePartnerProfile from "@/lib/swr/use-partner-profile";
 import useWorkspace from "@/lib/swr/use-workspace";
 import { PlanProps } from "@/lib/types";
 import { useLocalStorage } from "@dub/ui";
 import { fetcher } from "@dub/utils";
-import { endOfDay, startOfDay, subDays } from "date-fns";
 import { useParams, usePathname, useSearchParams } from "next/navigation";
 import {
   createContext,
@@ -34,6 +30,7 @@ import { toast } from "sonner";
 import useSWR from "swr";
 import { defaultConfig } from "swr/_internal";
 import { UpgradeRequiredToast } from "../shared/upgrade-required-toast";
+import { useAnalyticsQuery } from "./use-analytics-query";
 
 export interface AnalyticsDashboardProps {
   domain: string;
@@ -94,7 +91,7 @@ export default function AnalyticsProvider({
 }>) {
   const searchParams = useSearchParams();
   const pathname = usePathname();
-  const { id: workspaceId, slug, domains } = useWorkspace();
+  const { slug, domains } = useWorkspace();
 
   const [requiresUpgrade, setRequiresUpgrade] = useState(false);
 
@@ -107,52 +104,10 @@ export default function AnalyticsProvider({
   const partnerPage = partner?.id && programSlug ? true : false;
 
   const domainSlug = searchParams?.get("domain");
-  // key can be a query param (stats pages in app) or passed as a staticKey (shared analytics dashboards)
-  const key = searchParams?.get("key") || dashboardProps?.key;
 
   // Show conversion tabs/data for all dashboards except shared (unless explicitly set)
   const showConversions =
     !dashboardProps || dashboardProps?.showConversions ? true : false;
-
-  const tagIds = combineTagIds({
-    tagId: searchParams?.get("tagId"),
-    tagIds: searchParams?.get("tagIds")?.split(","),
-  })?.join(",");
-
-  const folderId = searchParams?.get("folderId") ?? undefined;
-
-  const customerId = searchParams?.get("customerId") ?? undefined;
-
-  // Default to last 24 hours
-  const { start, end } = useMemo(() => {
-    const hasRange = searchParams?.has("start") && searchParams?.has("end");
-
-    return {
-      start: hasRange
-        ? startOfDay(
-            new Date(searchParams?.get("start") || subDays(new Date(), 1)),
-          )
-        : undefined,
-
-      end: hasRange
-        ? endOfDay(new Date(searchParams?.get("end") || new Date()))
-        : undefined,
-    };
-  }, [searchParams?.get("start"), searchParams?.get("end")]);
-
-  const defaultInterval = partnerPage
-    ? DUB_PARTNERS_ANALYTICS_INTERVAL
-    : DUB_LINKS_ANALYTICS_INTERVAL;
-
-  // Only set interval if start and end are not provided
-  const interval =
-    start || end ? undefined : searchParams?.get("interval") ?? defaultInterval;
-
-  const selectedTab: EventType = useMemo(() => {
-    const event = searchParams.get("event");
-
-    return EVENT_TYPES.find((t) => t === event) ?? "clicks";
-  }, [searchParams.get("event")]);
 
   const [persistedSaleUnit, setPersistedSaleUnit] =
     useLocalStorage<AnalyticsSaleUnit>(`analytics-sale-unit`, "saleAmount");
@@ -217,7 +172,7 @@ export default function AnalyticsProvider({
       return {
         basePath: "",
         baseApiPath: "",
-        domain: "",
+        domain: "", // TODO [refactor]
       };
     }
   }, [
@@ -229,68 +184,33 @@ export default function AnalyticsProvider({
     partner?.id,
     programSlug,
     domainSlug,
-    key,
-    selectedTab,
   ]);
 
-  /*
-    If explicitly set, use the value
-    If not set:
-      - Show root domain links if:
+  const { queryString, key, start, end, interval, tagIds, selectedTab } =
+    useAnalyticsQuery({
+      domain: domain ?? undefined,
+      defaultKey: dashboardProps?.key,
+      defaultInterval: partnerPage
+        ? DUB_PARTNERS_ANALYTICS_INTERVAL
+        : DUB_LINKS_ANALYTICS_INTERVAL,
+
+      /*
+      If not explicitly set, show root domain links if:
         - it's filtered by a link, or
         - the workspace has more than 50 domains
         - is admin page
         - is filtered by a folder or tag
-      - Otherwise, hide root domain links
-  */
-  const root = searchParams.get("root")
-    ? searchParams.get("root") === "true"
-    : (domain && key) ||
+      Otherwise, hide root domain links
+    */
+      defaultRoot: ({ key, folderId, tagIds }) =>
+        (domain && key) ||
         (domains && domains.length > 50) ||
         adminPage ||
         folderId ||
         tagIds
-      ? undefined
-      : "false";
-
-  const queryString = useMemo(() => {
-    const availableFilterParams = VALID_ANALYTICS_FILTERS.reduce(
-      (acc, filter) => ({
-        ...acc,
-        ...(searchParams?.get(filter) && {
-          [filter]: searchParams.get(filter),
-        }),
-      }),
-      {},
-    );
-    return new URLSearchParams({
-      ...availableFilterParams,
-      ...(workspaceId && { workspaceId }),
-      ...(domain && { domain }),
-      ...(key && { key }),
-      ...(start &&
-        end && { start: start.toISOString(), end: end.toISOString() }),
-      ...(interval && { interval }),
-      ...(tagIds && { tagIds }),
-      ...(root && { root: root.toString() }),
-      event: selectedTab,
-      ...(folderId && { folderId }),
-      ...(customerId && { customerId }),
-      timezone: Intl.DateTimeFormat().resolvedOptions().timeZone,
-    }).toString();
-  }, [
-    workspaceId,
-    domain,
-    key,
-    searchParams,
-    start,
-    end,
-    tagIds,
-    root,
-    selectedTab,
-    folderId,
-    customerId,
-  ]);
+          ? true
+          : false,
+    });
 
   // Reset requiresUpgrade when query changes
   useEffect(() => setRequiresUpgrade(false), [queryString]);
